Move spinner to calling clients
