+++ client-common/Cargo.toml
@@ -12,7 +12,6 @@ edition = "2021"
 [dependencies]
 base64 = "0.13.0"
 bs58 = "0.4.0"
-indicatif = "0.17.0"
 jsonrpc-core = "18.0.0"
 reqwest = { version = "0.11.11", default-features = false, features = ["blocking", "brotli", "deflate", "gzip", "rustls-tls", "json"] }
 semver = "1.0.13"
