+++ client/src/lib.rs
@@ -15,7 +15,6 @@ pub mod rpc_filter;
 pub mod rpc_request;
 pub mod rpc_response;
 pub mod rpc_sender;
-pub mod spinner;
 pub mod thin_client;
 pub mod tpu_client;
 pub mod tpu_connection;
