--- client/src/spinner.rs
@@ -1,3 +1,0 @@
-//! Spinner creator
-
-pub use solana_client_common::spinner::*;
