+++ rpc-client/Cargo.toml
@@ -14,6 +14,7 @@ async-trait = "0.1.57"
 base64 = "0.13.0"
 bincode = "1.3.3"
 bs58 = "0.4.0"
+indicatif = "0.17.0"
 log = "0.4.17"
 reqwest = { version = "0.11.11", default-features = false, features = ["blocking", "brotli", "deflate", "gzip", "rustls-tls", "json"] }
 semver = "1.0.13"
