+++ tpu-client/Cargo.toml
@@ -17,6 +17,7 @@ enum_dispatch = "0.3.8"
 futures = "0.3"
 futures-util = "0.3.21"
 indexmap = "1.9.1"
+indicatif = "0.17.0"
 itertools = "0.10.2"
 lazy_static = "1.4.0"
 log = "0.4.17"
